requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
starlette==0.36.3
sqlalchemy==2.0.23
vaderSentiment==3.3.2
python-multipart==0.0.6
pydantic==2.4.2
alembic==1.13.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
transformers==4.37.2
torch==2.2.0
redis==5.0.1

